+++ device/util.go
@@ -3,9 +3,9 @@ package device
 import (
 	"GADS/util"
 	"context"
+	"fmt"
 	"time"
 
-	log "github.com/sirupsen/logrus"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo/options"
 )
@@ -20,14 +20,15 @@ func GetLatestDBDevices() {
 	for {
 		cursor, err := collection.Find(context.Background(), bson.D{{}}, options.Find())
 		if err != nil {
-			log.Fatal(err)
+			util.ErrorLog("gads-ui", "get_devices", fmt.Sprintf("Could not get db cursor when trying to get latest device info from db - %s", err))
 		}
 
 		if err := cursor.All(util.MongoCtx(), &latestDevices); err != nil {
-			log.Fatal(err)
+			util.ErrorLog("gads-ui", "get_devices", fmt.Sprintf("Could not get devices latest info from db cursor - %s", err))
 		}
+
 		if err := cursor.Err(); err != nil {
-			log.Fatal(err)
+			util.ErrorLog("gads-ui", "get_devices", fmt.Sprintf("Encountered db cursor error - %s", err))
 		}
 
 		cursor.Close(util.MongoCtx())
